instance-diffusion/requirements.txt
absl-py==2.1.0
accelerate==0.29.3
antlr4-python3-runtime==4.9.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clip==0.2.0
contourpy==1.1.1
cycler==0.12.1
diffusers==0.27.2
einops==0.7.0
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.2
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
packaging==24.0
pandas==2.0.3
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==0.9.16
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
streamlit==1.35.0